transcricoes-fireflies/requirements.txt
boto3
requests
fpdf2 2.8.1
urllib3
datetime
python-dotenv